Break flare26 js file into separate components
